Cleanup: USD: various non-functional changes
- Use const in more places
- Use more appropriate headers and forward decls elsewhere
- Remove one instance of variable shadowing
- Rename an internal rna enum list and the USD operator property object
  to follow surrounding code and naming guidelines

Pull Request: https://projects.blender.org/blender/blender/pulls/136112

+++ source/blender/io/usd/intern/usd_hash_types.hh
@@ -28,7 +28,7 @@ template<> struct DefaultHash<pxr::TfToken> {
 template<> struct DefaultHash<pxr::SdfPath> {
   uint64_t operator()(const pxr::SdfPath &value) const
   {
-    return (uint64_t)value.GetHash();
+    return value.GetHash();
   }
 };
 }  // namespace blender
